backend/requirements.txt
absl-py==2.4.0
astunparse==1.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
flatbuffers==25.12.19
gast==0.4.0
google-auth==2.49.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.78.0
h5py==3.14.0
idna==3.11
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==1.26.0
oauthlib==3.3.1
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pillow==12.1.1
protobuf==3.19.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
requests==2.33.0
requests-oauthlib==2.0.0
rich==14.3.3
six==1.17.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.7
wrapt==2.1.2
